production

+++ src/basket/basket.service.ts
@@ -1,13 +1,26 @@
 import {ObjectId} from 'mongodb';
-import {BasketModel} from './basket.model';
+import {Basket, BasketModel} from './basket.model';
 import {BasketProduct, BasketProductModel} from './basketProduct.model';
 import {BasketProductDto} from "./basketProduct.dto";
 import ProductService from "../product/product.service";
+import ApiError from "../exceptions/ApiError";
+import {BasketDto} from "./basket.dto";
 
 class BasketService {
 
     getBasketProductDto(basketProduct: BasketProduct) {
         return new BasketProductDto(basketProduct)
+    }
+
+
+    async getBasket(userId: string) {
+        const basket = await BasketModel.findOne<Basket>({
+            userId: new ObjectId(userId),
+        });
+        if (!basket) {
+            throw ApiError.badRequest('not found')
+        }
+        return new BasketDto(basket)
     }
 
     async createBasket(id: ObjectId) {
@@ -24,30 +37,39 @@ class BasketService {
 
         if (productInDb) {
             productInDb.count += 1;
-            return productInDb.save();
+            await productInDb.save();
+            return this.getBasketProductDto(productInDb);
         } else {
             const product = await BasketProductModel.create({
                 basketId,
                 productId,
             });
-
-            const basketProductDto = this.getBasketProductDto(product)
-            return basketProductDto;
+            return this.getBasketProductDto(product);
         }
     }
 
-    async deleteProduct(id: string) {
-        const product = await BasketProductModel.findByIdAndDelete(id);
+    async deleteProduct(basketId: string, productId: string) {
+
+        const product = await BasketProductModel.findOneAndDelete({$and: [{basketId: basketId}, {productId}]});
         return product;
     }
 
-    async updateCountProduct(productId: string, count: number) {
-        const updatedProduct = await BasketProductModel.findByIdAndUpdate(
-            productId,
-            // @ts-ignore
-            {$set: count},
+    async deleteAllProduct(basketId: string) {
+        const products = await BasketProductModel.deleteMany({basketId});
+        return products;
+
+    }
+
+
+    async updateCountProduct(basketId: string, productId: string, count: number) {
+        const updatedProduct = await BasketProductModel.findOneAndUpdate(
+            {$and: [{basketId}, {productId}]},
+            {$set: {count: count}},
             {new: true},
         );
+        if (!updatedProduct) {
+            throw ApiError.badRequest('Count don\'t updated')
+        }
         if (count === 0) {
             const product = await BasketProductModel.findByIdAndDelete(productId);
             return product;
@@ -77,6 +99,7 @@ class BasketService {
 
         return processProducts();
     }
+
 }
 
 export default new BasketService();
